+++ cmp/viewmanager/ViewManagerModel.ts
@@ -19,7 +19,7 @@ import type {ViewManagerProvider, ReactionSpec} from '@xh/hoist/core';
 import {genDisplayName} from '@xh/hoist/data';
 import {fmtDateTime} from '@xh/hoist/format';
 import {action, bindable, makeObservable, observable, comparer, runInAction} from '@xh/hoist/mobx';
-import {olderThan, SECONDS} from '@xh/hoist/utils/datetime';
+import {SECONDS} from '@xh/hoist/utils/datetime';
 import {executeIfFunction, pluralize, throwIf} from '@xh/hoist/utils/js';
 import {find, isEqual, isNil, isNull, isObject, isUndefined, lowerCase, uniqBy} from 'lodash';
 import {ReactNode} from 'react';
@@ -92,9 +92,8 @@ export interface ViewManagerConfig {
 
     /**
      * Delay (in ms) to wait after state has been set on associated components before listening for
-     * further state changes. The long default wait 1000ms is intended to avoid a false positive
-     * dirty indicator when linking to complex components such as dashboards or grids that can
-     * report immediate changes to state due to internal processing or rendering.
+     * further state changes. Can be overridden by `PersistOptions.settleTime` on individual
+     * `PersistenceProvider` instances.
      */
     settleTime?: number;
 
@@ -233,9 +232,6 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
     /** Data access for persisting views. */
     private dataAccess: DataAccess<T>;
 
-    /** Last time changes were pushed to linked persistence providers */
-    private lastPushed: number = null;
-
     //---------------
     // Getters
     //---------------
@@ -304,7 +300,7 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
         enableGlobal = true,
         enableSharing = true,
         preserveUnsavedChanges = true,
-        settleTime = 1000,
+        settleTime,
         initialViewSpec = null
     }: ViewManagerConfig) {
         super();
@@ -430,10 +426,7 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
 
     @action
     setValue(value: Partial<T>) {
-        const {view, pendingValue, lastPushed, settleTime} = this;
-        if (!pendingValue && settleTime && !olderThan(lastPushed, settleTime)) {
-            return;
-        }
+        const {view, pendingValue} = this;
 
         value = this.cleanState(value);
         if (!isEqual(value, view.value)) {
@@ -445,11 +438,6 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
         } else {
             this.pendingValue = null;
         }
-    }
-
-    /** @internal Called by ViewManagerProvider. */
-    noteStatePushed() {
-        this.lastPushed = Date.now();
     }
 
     //------------------