+++ core/persist/provider/ViewManagerProvider.ts
@@ -5,7 +5,6 @@
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
 
-import {Persistable} from '@xh/hoist/core';
 import {throwIf} from '@xh/hoist/utils/js';
 import {pull} from 'lodash';
 import {PersistenceProvider, PersistenceProviderConfig} from '../PersistenceProvider';
@@ -15,9 +14,16 @@ export class ViewManagerProvider<S> extends PersistenceProvider<S> {
     readonly viewManagerModel: ViewManagerModel;
 
     constructor(cfg: PersistenceProviderConfig<S>) {
-        super(cfg);
-        const {viewManagerModel} = cfg.persistOptions;
+        const {persistOptions} = cfg,
+            {viewManagerModel} = persistOptions;
         throwIf(!viewManagerModel, `ViewManagerProvider requires a 'viewManagerModel'.`);
+        super({
+            ...cfg,
+            persistOptions: {
+                ...persistOptions,
+                settleTime: persistOptions.settleTime ?? viewManagerModel.settleTime
+            }
+        });
         this.viewManagerModel = viewManagerModel;
         viewManagerModel.providers.push(this);
     }
@@ -25,17 +31,11 @@ export class ViewManagerProvider<S> extends PersistenceProvider<S> {
     pushStateToTarget() {
         const state = this.read();
         this.target.setPersistableState(state ? state : this.defaultState);
-        this.viewManagerModel.noteStatePushed();
     }
 
     //----------------
     // Implementation
     //----------------
-    override bindToTarget(target: Persistable<S>) {
-        super.bindToTarget(target);
-        this.viewManagerModel.noteStatePushed();
-    }
-
     override readRaw() {
         return this.viewManagerModel.getValue();
     }