+++ desktop/cmp/panel/PanelModel.ts
@@ -278,6 +278,7 @@ export class PanelModel extends HoistModel implements Persistable<PanelPersistSt
             PersistenceProvider.create({
                 persistOptions: {
                     path: 'panel',
+                    settleTime: 300,
                     ...persistWith
                 },
                 target: this